backend/core/pyproject.toml
[project]
name = "tank-backend"
version = "0.1.0"
description = "Tank Voice Assistant - Backend API Server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "httpx>=0.24.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "requests>=2.32.5",
    "beautifulsoup4>=4.13.5",
    "sounddevice>=0.5.5",
    "silero-vad>=4.0.0",
    "onnxruntime>=1.16.1",
    "faster-whisper>=1.0.0",
    "sherpa-onnx>=1.12.23",
    "fastapi>=0.129.0",
    "uvicorn>=0.40.0",
    "websockets>=12.0",
    "python-multipart>=0.0.22",
    "sherpa-onnx-core>=1.12.25",
    "tank-contracts",
    "docker>=7.0.0",
    "tiktoken>=0.7.0",
]

[project.scripts]
tank-backend = "tank_backend.main:main"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.lint.isort]
known-first-party = ["tank_backend"]

[tool.uv.sources]
tank-contracts = { workspace = true }
